refactor: separate the household items types from the types file
